Content Permission & Service Refactoring: A number of big changes to content services including new permission services. This is the first pass with detailed testing on blogs in ecommerce. Tested low level screens as well.
This is the first pass, there are still a number of references to old permission checking, some services need to be refactored, some replaced with better practices. This is a work in progress. 

This first pass includes the removal of special BLOG_ roles and status. Replaced with a generic CONTENT_ and CTNT_. Also adds new ECAs to base createContent service which creates OWNER role as well as associations when passed a contentIdTo or contentIdFrom (and TYPE). This will probably replace the persistContentAndAssoc service which is used in many places. Some have been updated in the revision, others will follow.

Most code which calls the older permission services are found in .java files. Almost all simple methods have been converted to have the permission in the service def and the code removed from the implementation. Java methods still need to be addressed.

The older permission service will output a warning in logs with a stack trace to help track down places where this service is still being called. This is not an error, but may show in logs for the time being.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/content@509321 13f79535-47bb-0310-9956-ffa450edef68