More fixes in the forum feature in the content component. I also added a few comments to indicate what different fields are used for.
The eCommerce component organized forum data one way, and the content component organized it another way. After discussion on the dev ml, I determined that the content component was storing data incorrectly.

This commit is not backwards compatible with existing forum messages. Existing forum messages will have to have the contents of ContentAssoc.contentId and ContentAssoc.contentIdTo swapped before updating to this revision.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/content@569568 13f79535-47bb-0310-9956-ffa450edef68