Fix a bug reported by James McGill on user ML :
content/control/EditSurveyResponse puts escaped HTML on the page instead of
the rendered survey.
applications/content/webapp/content/survey/EditSurveyResponse.ftl  has
${surveyString}    but
that apparently escapes the page.   ${surveyString?html} just adds another
level of escaping, and
I couldn't see how to "unescape" from here.
It looks to me like there really should be a call to surveyWrapper.render()
or something like that --
or is it the intended behavior to show the  HTML output instead of the
rendered survey?


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/content@815687 13f79535-47bb-0310-9956-ffa450edef68