Applied patch from jira issue - OFBIZ-3923 - Bug in updating local file content.
attachUploadToDataResource (member service of updateContentAndUploadedFile service group) service calls saveLocalFileDateResource simple method to save the local file and attach it to data resource. The case where updating a local file content but not uploading a new file is being handled but doesn't work since on finding DataResource object "isUpdate" flag is not set to "Y".

Here is the patch for the fix. Also fixed typo in simple method name, saveLocalFileDateResource => saveLocalFileDataResource.

Thanks Mridul for the contribution.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/content@993344 13f79535-47bb-0310-9956-ffa450edef68