Deprecate the local getArg/getWrappedObject wrappers in all transforms.
Several ftl transforms had local methods, to that wrapped the same named
method in FreeMarkerWorker.  Near as I can tell, this was to reduce
local typing, by not having to always type FreeMarkerWorker.  Having
these local methods increases the permgen memory use(by increasing the
size of the bytecode).  Also, it makes it more difficult to
cross-reference; when the getArg method is seen by a developer, is it a
local method, or a static import?  Using the fully qualified variant
makes it easier to understand what is going on.

The replacement is to use the fully qualified versions in
FreeMarkerWorker.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/content@1044231 13f79535-47bb-0310-9956-ffa450edef68