A patch from René Scheibe "Incorrect CSS path for elRTE editor." (https://issues.apache.org/jira/browse/OFBIZ-4157) - OFBIZ-4157
The path to elrte-inner.css is incorrect when starting the editor.

The result is that in the WYSIWYG view eg. div and tables are not surrounded with a small blue border.
The problem is also visible in the logs: "Unknown request [css]; this request does not exist or cannot be called directly."

JLR: I checked this was the sole issue for elrte-inner.css

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/content@1065147 13f79535-47bb-0310-9956-ffa450edef68