A slightly modified patch from Supachai Chaima-ngua for <<Upload product image does not work.>> https://issues.apache.org/jira/browse/OFBIZ-6769
Upload product image dose not working because "updateDataResource" require out parameter name 'dataResourceId' but the service doesn't return by using entity-auto engine and upload image function need delegator but do not sent to upload image function.

jleroux: this was due to recent r1716271 so no needs to backport. The changes in ImageManagementServices were not operationally needed (it worked as is) but indeed better for a method like scaleImageMangementInAllSize to rely on DispatchContext rather than service context when called from a service. Since it's not a service I changed the scaleImageMangementInAllSize method to private. I also removed the useless "ImageManagementServices." prefix.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/content@1721066 13f79535-47bb-0310-9956-ffa450edef68