A patch from Ingo Wolfmayr for "Problem assigning contentassoc with different content assoc types" https://issues.apache.org/jira/browse/OFBIZ-7159
Go to: Content --> Content -> "WebStoreCONTENT" --> Association

Add an association:

ContentIdTo: CNTGIZMOS
ContentAssocTypeId: Alternate Locale
--> Create

add another via "create new association"
ContentIdTo: CNTGIZMOS
ContentAssocTypeId: Alternate Role
--> Create

The 2nd association is create with the ContentAssocTypeId of the first one created. You have to logout and login to clear that.

Cause: the selected ContentAssocTypeId does not change by selecting another value.
Solved it by removing to-scope="user"
I did look for reasons but could not find one
+ checked if everything is working
    create association works (first time)
    update works
    create further association works
I guess it is done to set the default value from the last selection, but it breaks functionality as you wont be able to change it.

jleroux: this was very old code (pre Apache era) which maybe worked before but does not longer. Anyway the old behaviour (if it worked) was very minor, so I see no problems removing it to have something working




git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/content@1746755 13f79535-47bb-0310-9956-ffa450edef68