Fix for: Freemarker's ignore_missing attribute of the #include directive was not working because of an issue in the OFBiz custom TemplateLoader for Freemarker templates. (OFBIZ-8292)
The OFBiz custom TemplateLoader now returns null if the resource is missing as 
required by the TemplateLoader specification.
Additional cleanups for unused methods in the FreeMarkerWorker class and some 
minor fine tuning; improved the way errors are rendered: now the full stack 
trace is not shown in the screen but only in the logs.

Thanks: Jacques Le Roux for the report.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/content@1761586 13f79535-47bb-0310-9956-ffa450edef68